feat: bdd unit tests added for httpTool example
